Refractored the Hue code to add interfaces
